LoginPasskey.dart
This is an early attempt at building authentication using Passkeys. At the time of writing, Passkeys were still 'catching on' and documentation was spurious at best. This code never made it to production, but is a good start based on Google, Apple, and Email authentication as a beginning structure, and has similarities to the LoginScreen file in this repo.